pyproject.toml
[tool.poetry]
name = "django"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django-test = "^0.4030"
bs4 = "^0.0.1"
requests = "^2.31.0"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
